Add files via upload
I made further changes to the edit grade functionality

+++ CHM2045.h
@@ -8,7 +8,7 @@ private:
 public:
 	void calcGpa();
 	void updateExam(int examNum, double score);
-	void updateFinal(int score);
+	void updateFinal(double score);
 	void updateHomework(int homeworkNum, double score);
 	void updateHittPoints(double score);
 	void printAll();
